requirements.txt
discord.py
Pillow
colorama
termcolor